cleanup test structure
