+++ backend/packages/cffdrs-vec/src/cffdrs_vec/tests/_glc_x_10_data.py
@@ -1,7 +1,7 @@
 """
 Shared loading/computation logic for the GLC-X-10 (Wotton et al. 2009) published FBP System test
-cases, used by test_fbp_glc_x_10.py (reads the vendored CSVs directly) and
-test_fbp_glc_x_10_rasters.py (reads the equivalent GeoTIFF fixtures generated by
+cases, used by test_fbp_glc_x_10.py's test_fbp_glc_x_10 (reads the vendored CSVs directly) and
+test_fbp_glc_x_10_rasters (reads the equivalent GeoTIFF fixtures generated by
 generate_glc_x_10_rasters.py from those same CSVs), both parametrized over both sources. See
 https://github.com/bcgov/wps/issues/4893.
 
@@ -360,7 +360,7 @@ class GLCX10Source(ABC):
 
     def to_fbp_inputs(self) -> list[FBPInput]:
         """One FBPInput per case, in the same order as load()'s arrays - used by
-        test_fbp_e2e_vectorized.py to call cffdrs's own, real fire_behaviour_prediction() once per
+        test_fbp_matches_cffdrs.py to call cffdrs's own, real fire_behaviour_prediction() once per
         case, as the ground truth fbp.vectorized_fire_behaviour_prediction gets checked against.
         Doesn't go through any of cffdrs_vec's own vectorized/jitted code at all.
