+++ backend/packages/cffdrs-vec/src/cffdrs_vec/tests/test_fbp_glc_x_10.py
@@ -11,8 +11,13 @@ provenance of each):
   with no R package in the chain at all.
 If these two ever disagree, the R-sourced CSVs and the paper have actually diverged and that's
 worth investigating; both staying green is the cross-validation working as intended.
-test_fbp_glc_x_10_rasters.py runs the same two sources through GeoTIFF rasters instead, to prove
-the array pipeline also works end-to-end through real raster I/O.
+
+test_fbp_glc_x_10 reads inputs straight from the CSVs; test_fbp_glc_x_10_rasters runs the same
+two sources through the equivalent GeoTIFF fixtures under fixtures/glc_x_10/ and
+fixtures/glc_x_10_paper/ (generated by generate_glc_x_10_rasters.py, via GLCX10Source.
+load_raster_inputs()) instead, to prove the array pipeline also works end-to-end through real
+GDAL raster reads (float32 storage, WPSDataset's nodata handling) and not just plain float64
+arrays parsed straight from CSV.
 
 ROS, HFI and CFB use a looser tolerance for "paper" than "r_package" (GLCX10Source.rtol_ros_hfi/
 atol_cfb): cffdrs_vec (like the current R cffdrs package) is consistently ~0.3-0.6% off the
@@ -20,7 +25,7 @@ paper's own published values for these three outputs specifically, while SFC/TFC
 match tightly for both sources. That's consistent with refinements made to the FBP
 rate-of-spread/CFB equations in the years since this 2009 paper was published.
 vectorized_fire_behaviour_prediction was already cross-checked against cffdrs's own
-fire_behaviour_prediction() directly (test_fbp_e2e_vectorized.py), so the goal here is catching
+fire_behaviour_prediction() directly (test_fbp_matches_cffdrs.py), so the goal here is catching
 gross regressions against the paper, not bit-for-bit reproduction of a since-superseded formula.
 """
 
@@ -74,4 +79,89 @@ def test_fbp_glc_x_10(source):
     np.testing.assert_allclose(result.tfc, expected.tfc, rtol=1e-3, err_msg="TFC")
     np.testing.assert_allclose(result.raz, expected.raz, atol=0.1, err_msg="RAZ")
     fire_type = [FIRE_TYPE_BY_FD_CODE_INT[c] for c in result.fd_code]
+    assert fire_type == expected.fire_type
+
+
+@pytest.mark.parametrize("source", GLC_X_10_SOURCES, ids=lambda s: s.name)
+def test_raster_inputs_match_csv_inputs(source):
+    """Sanity check that generate_glc_x_10_rasters.py's output hasn't drifted from the CSVs it
+    was generated from - this would fail if the CSVs were updated without regenerating rasters.
+    """
+    # raster_inputs comes back 2-D (eg. (1, 20), same shape a real raster grid would have);
+    # csv_inputs is flat, one value per case - flatten the raster side to compare them.
+    raster_inputs = source.load_raster_inputs()
+    csv_inputs, _expected = source.load()
+    np.testing.assert_allclose(raster_inputs.fuel_type_codes.flatten(), csv_inputs.fuel_type_codes)
+    np.testing.assert_allclose(raster_inputs.ffmc.flatten(), csv_inputs.ffmc)
+    np.testing.assert_allclose(raster_inputs.bui.flatten(), csv_inputs.bui)
+    np.testing.assert_allclose(raster_inputs.cbh.flatten(), csv_inputs.cbh)
+    np.testing.assert_allclose(raster_inputs.cfl.flatten(), csv_inputs.cfl)
+
+
+@pytest.mark.parametrize("source", GLC_X_10_SOURCES, ids=lambda s: s.name)
+def test_fbp_glc_x_10_rasters(source):
+    inputs = source.load_raster_inputs()
+    shape = inputs.fuel_type_codes.shape
+    zeros = np.zeros(shape)
+    int_zeros = np.zeros(shape, dtype=np.int64)
+
+    result = fbp.vectorized_fire_behaviour_prediction(
+        inputs.fuel_type_codes,
+        inputs.ffmc,
+        inputs.bui,
+        inputs.ws,
+        inputs.wd_rad,
+        inputs.gs,
+        inputs.aspect_rad,
+        inputs.pc,
+        inputs.pdf,
+        inputs.cc,
+        inputs.gfl,
+        inputs.cbh,
+        inputs.cfl,
+        zeros,  # fmc - all 20 published cases leave it unset, so it's always derived internally
+        zeros,  # isi - all 20 published cases leave it unset, so it's always derived internally
+        inputs.lat,
+        inputs.lon,
+        inputs.elv,
+        inputs.dj,
+        inputs.d0,
+        zeros,  # sd - no stand density data in GLC-X-10
+        zeros,  # sh - no stand height data in GLC-X-10
+        np.ones(shape),  # hr
+        zeros,  # theta_rad - unused by Primary output
+        int_zeros,  # accel
+        np.ones(shape, dtype=np.int64),  # buieff - use raw BUI, not effective BUI
+    )
+
+    np.testing.assert_allclose(
+        result.ros, source.read_raster_output("ros"), rtol=source.rtol_ros_hfi, err_msg="ROS"
+    )
+    np.testing.assert_allclose(
+        result.hfi, source.read_raster_output("hfi"), rtol=source.rtol_ros_hfi, err_msg="HFI"
+    )
+    np.testing.assert_allclose(
+        result.cfb, source.read_raster_output("cfb"), atol=source.atol_cfb, err_msg="CFB"
+    )
+    np.testing.assert_allclose(
+        result.sfc, source.read_raster_output("sfc"), rtol=1e-3, err_msg="SFC"
+    )
+    np.testing.assert_allclose(
+        result.tfc, source.read_raster_output("tfc"), rtol=1e-3, err_msg="TFC"
+    )
+
+    # _fire_behaviour_prediction wraps the angle with an exact `raz_deg == 360 -> 0` check. Float32
+    # raster storage rounds just enough that one case can land on 360.00001 instead of 0 - the same
+    # angle, but not an exact match - so compare the circular (mod-360) distance instead.
+    expected_raz = source.read_raster_output("raz")
+    circular_diff = np.abs((result.raz - expected_raz + 180) % 360 - 180)
+    np.testing.assert_allclose(circular_diff, np.zeros_like(circular_diff), atol=0.1, err_msg="RAZ")
+
+    # fire_type isn't stored in a raster (it's derived from cfb, not read directly) - source.load()
+    # already builds the same by-id lookup against the output CSV that this needs, in the same
+    # input-row/pixel order, so there's nothing raster-specific left to redo here. fd_code comes
+    # back 2-D (same shape as the raster-sourced inputs, eg. (1, 20)), so flatten it to compare
+    # against expected.fire_type's flat, one-value-per-case list.
+    _, expected = source.load()
+    fire_type = [FIRE_TYPE_BY_FD_CODE_INT[c] for c in result.fd_code.flatten()]
     assert fire_type == expected.fire_type
